Experiment with Lua syntax
Using Lua as a scripting language seems like a great idea. If I invent
my own DSL it will be limited at best, and at some point you want a full
scripting language ... Or not? On the other hand, writing everything
directly in Lua is not convenient either. In particular, starting a new
slide should be painless, and unit suffixes such as "em" would be nice.

Some form of looping would be nice, but apart from that, my requirements
for *scripting* are quite basic. Expressions for basic arithmetic, but
that is not too hard. Apart from that, I think I mostly need assignment
to local variables, and functions to call. There is little need for
control flow when programming graphics.