Add commentless tweaked syntax example
The default LALRPOP lexer just splits everything at whitespace, so there
is no way to ignore comments. Including them in the grammar is not an
option either, because parsing rules cannot be no-ops as far as I know.
Well maybe it is possible, but it will not be pretty.

I'll write my own lexer. But first things first, for now not having
comments will do.