Make numbers dimensionful
This gets rid of the "len" value and coordinate. There are only numbers,
but numbers carry a length unit, characterized by the dimension. So now
I can also represent areas or inverse length. This is probably not going
to be useful, but it is simpler because there are less special cases.
Furthermore, being able to represent intermediate values that don't have
length or no units can be useful. For instance:

    (4.0 / 1w) * 1w

Should evaluate to a dimensionless number. Previously it would have
failed, because inverse length could not be represented. Now this works.