Implement function call for user-defined functions
Because the body gets evaluated in the context of the call, unexpected
things can happen though. For instance, the following code does not
work:

    f = function(a) { return function(b) { return a + b } }
    x = f(1)(2)

This will fail, because when the second function is called, 'a' is not
in scope. The call is evaluated in the global scope, where only 'f'
exists at that point. So basically, functions can assume nothing about
what values are in scope. This could be problematic and
counter-intuitive on the one hand. On the other hand, if a function
generates some text, I want it to respect the 'font_size' variable at
the call site, not the one that happened to be in effect where the
function was defined. Without this, I'd need to pass in far too much
state for every call. At least, that is my feeling right now. For
example:

    font_size = 12pt
    red_text = function(t)
    {
      color = #ff0000
      at (0w, 0w) put t
    }
    font_size = 24pt
    at (1w, 1w) put red_text("hi")
    font_size = 36pt
    at (1w, 2w) put red_text("there")

I think that I want this to produce text of two sizes: 24pt and 36pt,
not text of 12pt. We'll see how this goes.