Add 't' function for making text frames
There should likely be sugar for this later, because single-letter
functions do not scale. Longer functions work, but that is too
intrusive. But once I get to italics, this is going to be a problem:
you don't want a builtin function 'i', because it is going to clash with
variables (although there are no loops yet).

For now though, this function does not do anything. It is supposed to
typeset a string and produce a frame with the right text element, but at
the moment my frames do not contain any graphic elements at all, and I
don't have any Harfbuzz integration either.