Add '~', the adjoin operator
I previously wanted to use '+', but that would overload '+' in a sense.
Currently it only adds vectors (scalars or coords). Making it place
frames next to eachother would complect addition with adjoining frames.

Adjoining in combination with prefixes is what will make text formatting
ergonomic:

    t"Hello, " ~ i"world"

Compare that to html:

    <p>Hello, <i>world</i></p>

Or to LaTeX:

    Hello, \textit{world}

Or even Markdown:

    Hello, *world*  -- or -- Hello, _world_

Note that I am intentionally using the *markup* tags in these examples,
not the semantic tags (<em> and \emph). Pris will be about markup. It
has nothing to do with semantics.