Implement string concatenation
I hesitated a while about using '++' or '+' for string concatenation.
From the point of complecting things, it is best to keep addition and
concatenation separate. But on the other hand, many languages use '+'
for string concatenation already, so respect the principle of least
surprise. Though '++' is not that weird either.

In the end, having type safe, non-overloaded operators is mostly useful
in statically typed languages, I think. And when you mix operands, the
destinction can be clarifying. But in this case, splitting them out has
no advantage over keeping them the same, apart from not accidentally
adding numbers when you are certain that concatenation is meant. But
that is a rare mistake to make, if it is made at all. On the other hand,
I expect many people to try and add strings using '+', so it makes sense
to have that. Even there would be a note attached to the error that
suggests using '++'. That is an extra step, an unnecessary step. So
single plus it is.