Resolve last error in example document
It can now be fully compiled/interpreted, whatever you like to call it,
by the application. There are stubs of course. No graphics are placed
yet. But these are marked with to dos, and in most cases it is roughly
clear what must be done there.

There are also still a few rough edges on the syntax. I would like to
have prefixes instead of the t() function. There should be a dedicated
lexer, in order to support comments. But for now, this will do. I can
have a test document, and it compiles.