Add basic program that draws a pdf with Cairo
I want to try Skia as well for a fair comparison, but Cairo seems so
attractive. The API is straightforward, clear, reasonably
well-documented, and I could link to it directly from Rust.

Skia, on the other hand, needs to be built from source using an obscure
build system (GN is really nice, but the only way to obtain it as far as
I am aware, is to download a binary from somewhere). It is C++ so it is
hard to bind to from a different language. Maybe I'll just go with
Cairo.