Add basic Cairo pdf surface, link from Rust
The C ABI and Rust's FFI are nice. I don't even need to write C code, I
can just call it directly. Though there is a certain elegance to the raw
C code. Everything looks so clean and simple. Rust does not with its
camel case types and double colons and Egyptian brackets.

But anyway, let's try to get this minimal Cairo example that draws a
line working from Rust. And then couple my language to it.

Why write the Cairo interop manually if there exist crates for that?

 * I found three crates, but none of those support pdf surfaces.
 * Reinventing the wheel is fun.
 * Having less dependencies is a feature.
 * I am on a train without an internet connection at the moment.

Though I should probably contribute support for pdf surfaces to one of
the libraries and upstream it. But for now, this will do.