Draw graphics element in frames to pdf with Cairo
This is the first proof of concept! I have a domain specific language
for drawing graphics, a toy interpreter for it that produces frames with
graphics elements, and a driver that converts those graphics elements
into a pdf. From idea to lines on a pdf in 19 days. Hurray!

Now it is time to add some more properties (configurable stroke width
and color), and support for text using Harfbuzz. Also I need a proper
lexer with support for comments.

But for now I am happy! This project turned out to be quite
staightforward after all, especially after switching to Rust. Cairo is a
nice library, though I knew that already. (Skia probably is too, but
building it is not trivial, and neither is linking to it.)