Add roadmap
Well, "roadmap". It is more of a list of to-do things. The GitHub issue
tracker is really nice, but it is a form of vendor lock-in, and it
doesn't work offline. I'd like to be able to get stuff done on the
train. And we already have a great system for decentralized tracking of
resources, so let's use it.