Write my own Rust fontconfig wrapper
There are at least three direct fontconfig wrapper crates, and a few
more high-level crates. Yet all of them are either tedious to use or
broken. Instead of taking on all the dependencies, I wrote a small
module that wraps the C API. In the end it is not so much code anyway.