Work around bug in Fontconfig 2.8
So here is what I think was happening: in older versions of Fontconfig
(used on ancient Ubuntu 12.04 that Travis uses), FcFontMatch would not
overwrite its result argument if the match was succesful. My C program
worked even with this version, because I left the variable
uninitialized. It likely was 0 because it was on fresh stack, which
corresponds to the success value. And I did not check that result
anyway. On Rust, I conservatively set the result to "no match" before
calling FcFontMatch. On my system with a more recent Fontconfig this was
not a problem, but on Travis (and inside my Vagrant box), with the older
Fontconfig, the result would not be set to "match", even when a match
was found.

Setting the result to "match" to begin with solves this issue inside my
Vagrant box, and I expect it will solve the issue on Travis as well.