Pass around font map in interpreter
At this point one might wonder, would it make sense to put the font map,
and a stack of context frames, in a struct, and use methods on that
instead of functions? Maybe. But there still is a clear distinction
between evaluation which does not modify the frame, and evaluation which
does. The font map needs to be mutable because it acts as a cache, but
it behaves as an append-only immutable thing. Putting it in a data
structure would require all methods to take self as &mut, even though
most don't modify the frame. It could still make sense to put the font
map only in a struct. But for now, let's just do this, and see how far
it gets me.