Place images at the correct position
This adds ffi bindings for Cairo transformations, and uses them to
move the svg image. No scaling yet though.

On a side note, the Cairo API is beautiful really. Nice and clean,
plain C. There is a certain elegance to it. It is also simple; not
much that can be used in the wrong way, and everything is obvious
and straightforward.