Add OS X to the build matrix
Being cross-platform is not really a goal for now -- especially Windows
support. But all the pieces are there, and at least on OS X it should be
possible to install all dependencies somewhat easily. So let's try
whether this works. If it does, it would make the project more
contributor-friendly.