Add support for typesetting multiline strings
Adding a snippet of code or even a basic bullet list is quite tedious
without this; you have to manually compute the y offsets. Still though,
big blocks of text are not composable: what if you want a word in italic
in the middle, but also center align? For a single line, adjoin works
well. But for multiple lines it does not. Perhaps a better solution in
the long run is lists and something to set those with a certain spacing
and alignment.

For code snippets I still want to set multiline strings though, I don't
want to write that as a list of lines. But that is more of a syntax
issue. Multiline string literals are definitely required too.