Add pre-lexer, part of a custom lexer
A custom lexer is required to handle single-line comments, and the
non-greedy raw string literals marked by multi-character sequences. The
syntax for raw (multi-line) string literals will be triple dashes, as
opposed to the triple backticks that may be familiar from Markdown, or
Python's triple quotes. The reason is purely aesthetical: dashes are
centered on the line, so a block delineated by dashes looks balanced,
whereas this is not the case for quotes, and definitely not for
backticks. Compare

  snippet =
  ```
  <some code snippet>
  ```

Versus

  snippet =
  """
  <some code snippet>
  """

Versus

  snippet =
  ---
  <some code snippet>
  ---

The last option looks best, in my opinion. It is also a little less
dense than the double quotes.

The pre-lexer will be extended with a proper lexer that produces more
fine-grained tokens (splitting on whitespace etc), and then it will need
to be integrated into the parser, in favor of Lalrpop's generated lexer.