Derive Debug for a number of types
This is required to unwrap a Result with my error type in the tests, as
it needs to print the error value if it is indeed an error. I did not
reuse the printing logic that Error has; the Debug trait is about debug
printing values, but the Error print() function is about pretty-printing
errors for the user.