Store token contents inline for lalrpop
It is a bit unfortunate that we have to do this: the triple already
contains the start and end index and lalrpop already takes the input
string, so it could just slice it. Why does all the data need to be
duplicated? Oh well, I'll write my own parser some time ...

The code compiles now, but keyword tokens are not handled, so parsing is
expected to fail.