Print sha256sum of compiled examples on CI
This allows me to verify whether produced documents are truly
reproducible. Of course I would like them to be, and at least they are
on a single machine. (What this really means is that the process is
deterministic.) But it might be that different Cairo versions produce
different output, or that different versions of a font are installed
(which is embedded into the pdf), and this can cause different outputs.
In order to judge whether documents are truly reproducible, and whether
it would make sense to build this assumption into my tests, make Travis
CI print the sha256sums.