Update Windows compilation guide
I do think I have figured out how to build deterministically. It still
is not as pretty as it should be, but it suffices for now.

When pkg-config is started by Command::new in build.rs, rather than from
an MSYS2 shell, the environment variables are different -- paths are not
the unixy paths like in the shell, but they are the full windowsy paths.
This breaks pkg-config. It is possible to force PKG_CONFIG_PATH to be
"/usr/lib/pkgconfig:/usr/share/pkgconfig:/lib/pkgconfig" and that helped
for me, but then passing the -L flags to rustc failed for some reason.
Cargo was passing them properly, but rustc still failed to find the
libraries. Whereas with the -L flags in RUSTFLAGS, everything does work.
At this point I am glad that *anything* compiles, and I'll stop asking
questions.