Add "copy deps" step to MSYS build script
This command will gather all dlls that the executable (transitively)
depends on, so it can be used outside of an MSYS environment. All of
these files together make up about 23 MiB on my system currently. I am
able to execute pris.exe --help in a regular cmd window now. The problem
now is that fontconfig does not find "sans roman" when compiling the
examples.