Do a dry run on Appveyor
It might be a good idea to do this on Travis as well. But there it is
not so important; if the binary is built it probably works. And
compiling the examples tests it. But on Windows we have this DLL mess,
and I am still far away from doing anything that involves fonts on
Windows. So run pris --help instead.