Revert "Configure Git line endings on Appveyor"
This causes the build to fail because the initial clone fails, because
the directory is not empty. Even though before clone I am only setting
two global Git config variables, which should not create the directory.
This appears to be a bug in Appveyor. But for now those line endings are
not yet important, so let's remove this step again.

This reverts commit c8e0bc1c043e5747c02651bd2238da57195f79cf.