Add prototype for a 'glyph()' function
This function will expose access to a single glyph. Currently drawing
the glyphs works, but the bounding box is incorrect; the glyph width
needs to be extracted somewhere, and in this case Harfbuzz does not do
it.

Furthermore, the example uses a font which is not freely available, so
it cannot be used on CI, and it is a bad example because people will not
be able to reproduce it in general.