Add link to Gitter chatroom
The chatroom is intended for help with using Pris, because there is not
really a good place for that, and my docs are still mostly absent.
GitHub is great for development discussions, and some projects do use
the issue tracker as a support channel, but I am a bit ambivalent about
that. Stack Overflow is also great, but for established projects, not
for something new and experimental. Let's try a chatroom then.