Split frames into subframes
Previously a frame would directly contain elements. Now a frame contains
one or more subframes, and those contain elements. This will enable easy
animation.

This commit splits replaces the elements collection in a frame with a
subframe collection, where a subframe contains elements.

The tricky part now, is combining subframes: placing and adjoining them.
This has not been implemented yet, and hence the code does not compile
at this point.

Furthermore, it is not yet possible to create additional subframes.
Syntax will be added for that, but that should not be very hard.