Move interpreter state into structs
Previously, the (mutable, for caching) font map, immutable environment
(for expressions), and mutable frame (for statements) were all threaded
through the interpreter functions manually. This does not scale, because
every extra piece of information needs to touch all functions, like when
the font map was added.

This is a bit messy still. There are now circular dependencies between
modules, and lifetime annotations proliferate. This should be replaced
with something simpler, but for now this will suffice.

Placement has still not been fixed to support subframes, hence this does
not compile currently.