Fix placement after adding subframes
Every frame can now consist of subframes, and placing a frame will place
all the subframes in a zipped fashion, starting at the current subframe.
The "current subframe" is initialized to 0, and at the moment it never
changes. A pause command like Beamer's \pause will be added to do this.