Avoid recursion in grammar for function calls
This was pointed out by Antlr when I tried to write the grammar in
there. Interestingly I encountered the exact same issue in a different
project today (a PEG-based parser generator) which lead to infinite
recursion, so this seems like a good idea nonetheless.

The size of the parser generated by lalrpop is also slightly smaller.