Start working on hand-rolled parser
My main problem with Lalrpop is that it is ridiculously slow to compile,
and it pulls in too many dependencies. It complicates the build. So I
would rather write the parser by hand then, even if the parser itself is
not as efficient. Furthermore, I can generate better error messages this
way.