Derive PartialEq for ast nodes
Also Eq wherever possible. This is required to be able to run unit
tests: I need to be able to compare actual output to expected output. It
is a bit of a hassle already, because the types do not implement Debug,
only my own pretty-printing trait. But maye I should add a Debug impl
for anything that implements Print.