Simplify parser by keeping cursor in state
The original idea behind returning it, was that it would be possible to
backtrack. But it turns out that my grammar is such that by looking
ahead 1 token, it is always possible to decide how to continue to parse.
There is no need for backtracking. So we can eagerly consume all tokens
and store the cursor inline in the parser.