Remove dependency on Lalrpop
This will cut down the compile time tremendously, especially for clean
builds (Lalrpop is the heaviest crate in the dependency graph, not
counting its transitive dependencies), but also for rebuilds, as it does
not have to regenerate the parser every time. The build script shouldn't
have done that in the first place, but it did anyway.

I will make a more detailed compile time comparison when merging the
parser work.