Implement painting of png images
Png images can now be used and displayed, but their size is not yet
computed properly. I was trying to use libpng to load the image to get
the size, but that is incredibly painful. Fortunately, the dimensions of
the image are a few bytes at a fixed offset from the beginning of the
file, so I can just read those manually.