Add test for parser associativity
There is currently a bug, where binary operators associate right, while
they should associate left. This is particularly bad in an expression
like

  a - b + c

Which would be parsed as

  a - (b + c)

rather than

  (a - b) + c

This commit does not fix this issue, but it at least adds a test for it.
I recently read about Pratt parsers, that might be a good fit here.