Fix associativity in binop parser
As an additional advantage, this removes some of the recursion from the
parser, which makes it less likely to cause a stack overflow on huge
expressions. It is still possible, just harder.

This does not use the Pratt parsers, just a slight modification of the
current parser to fix the issue. It is only a little less elegant than
the original recursive solution.

This bug does show that I need a better test suite for the parser
though.