Add canvas_size variable
The canvas_size variable controls the size of the canvas (in pixels).
But as there are no absolute lengths in Pris, it really controls the
aspect ratio of the canvas.

Note, for now this breaks all w and h lengths, because these are still
relative to the previous hard-coded canvas size.