Start working out infix calls
One more realisation I had: "at" does not need to be a keyword, it could
be a built-in function, if all binary functions could be called with
infix notation. Then apart from at, I can also use "fit" as infix:

    image("foo.svg") fit (5em, 5em)

Rather than

    fit(image("foo.svg"), (5em, 5em))

The former reads quite natural to me, and is less noisy. It could also
be done for lines if "line" were a binary function, or perhaps
"line_to":

    (1em, 2em) line (3em, 3em)

Rather than

    line((2em, 1em)) at (1em, 2em)

Although lines are a bad example, because I probably want to use a list
of coordinates for arbitrary outlines in the future, something like

    line([(1em, 2em) (3em, 3em)])
    line([(1em, 2em) (3em, 3em) (0em, 5em)])

Anyway, adding infix function calls makes the grammar ambiguous if I
stick with a single token of lookahead:

    foo = bar frobnicate baz
                         ^

    foo = bar
    frobnicate =
               ^

One solution would be to look ahead in this case. If there is a '='
after an identifier, then the identifier is not an infix call. If there
is not, then the identifier must be an infix call. This is what I try in
this commit, although Bison can no longer help is now, and its error
message is not particularly insightful.

Another alternative would be to require an additional token before an
infix call, for examply by surrounding the identifier with backticks,
like in Haskell, or maybe by prefixing with ' or !.

A third alternative would be to add statement terminators, probably
semicolons (although newlines could work too, but I would prefer not to
go down the significant whitespace rule). This has the additional
advantage that it captures errors earlier, and it would lead to better
error messages in general, I expect. For example, code like this:

    put fill_rectangle((0.5w, 0.5w))
    at (0em, 0em) put circle(0.5w))

which is now parsed as

    put fill_rectangle((0.5w, 0.5w)) at (0em, 0em);
    put circle(0.5w));

would cause an error at the "at" token, if statements were ;-terminated:

    put fill_rectangle((0.5w, 0.5w));
    at (0em, 0em) put circle(0.5w));
       ^
       expected '=' here, if 'at' is no longer a keyword
    ^
    or unexpected keyword 'at' here, if it is a keyword

I am open to the semicolons, they are not *that* noisy, and css and TikZ
have them as well. The only thing that I find a bit ugly is terminating
function bodies with them when defining a function.

    center = function(frame)
    {
      ...
    };
     ^
     slightly ugly

I think I could live with it though, but then on the other hand, with
just a bit of lookahead there is no need for separators at all. I'm
still not sure what to do.