Add fill_circle builtin function
It is not yet entirely implemented. I need to add support for curves,
and add the right control points to the function.

I considered special-casing circles, as Cairo has support for arcs, so
it could draw them with more precision. But it will mean extra cases,
and I want to have curves anyway, so I might as well add curves now and
build circles on top, even if they are not perfect circles.