Add example for lists
I want to add lists, to be able to draw curves. I struggled with the
syntax a bit, but I think a ; as element terminator is the least bad
solution. Other options were:

 * Space as separator, but that conflicts with infix calls.
 * Comma as separator or terminator, but it looks ugly as suffix in a
   multi-line list.

Semicolon as a terminator also looks a bit awkward on a single line,
but I could make it optional.