Write down ideas on unifying space and time
I started out writing down this idea that we could use subframes, which
I was planning to have for a long time, to make slides first class.
Along the way, I noticed a lot of similarities between space and time,
and that one construct (such as move/pause, at/dealy, and the
anchor/current subframe) which is useful in one domain, is also useful
in another domain. And in fact, I did not realize until just now that I
really wanted these features (mainly "move", and the "put+move"). In
physics of course it has been common practice for a very long time to
put space and time on the same footing, so in that sense it is no
surprise that it works here too, but still, this feels like a discovery
to me.

Everything seems so obvious in hindsight. Handling of animation and
spatial placement will be uniform, which makes the design of Pris feel
much more "fundamental" in some way, and much less ad-hoc.

I am pleased :)