Drop Docopt dependency
The CLI is so simple, that Docopt, which in turn depends on Serde and
Regex, does not pull its weight. Docopt is nice, don't get me wrong, but
the mass of dependencies it brings in for a task as trivial as this (in
the case of Pris), is not worth the benefit.

Also, for packaging, distro's like Debian package every crate
individually (even though they are only source packages). Going from 26
to 7 deps is a welcome improvement there.

The changes to the main program are minimal; dropping Docopt does not
really come at the cost of complexity. Although I did have to drop the
--output flag (with value separated by either = or a space). That would
have complicated things. If Pris were more complex, I could use the CLI
parser from Tako, which is fully featured but still a lot lighter than
external libraries. But even that is overkill here. The --output flag
has been replaced with a positional argument.

On Rust 1.28.0:

Before:
  Number of Rust deps:         26
  Binary size:                 5367912 bytes
  Binary size (stripped):      1555192 bytes
  Clean build time (debug):    21.88s / 21.45s / 21.60s
  Clean build time (release):  44.29s / 45.95s / 42.99s

After:
  Number of Rust deps:         7
  Binary size:                 4425184 bytes            82% of before
  Binary size (stripped):       744136 bytes            47% of before
  Clean build time (debug):    5.76s / 5.79s / 5.78s    27% of before
  Clean build time (release):  9.43s / 8.92s / 9.06s    20% of before