Pin Rust version
I want to be able to check out an old commit and still have it build,
and pinning the toolchain helps with that. But I also want to test
against multiple releases on CI, to ensure compatibility, and to not
force this exact Rust version onto people who might want to build or
package Pris.