Make binary less chatty
At first I just wanted to add a carriage return between the frame
painting (which is what takes the longest, especially if you use
external images). But then even that seemed verbose, a proper Unix tool
is silent on success, but loud about failures. So make Pris silent on
success.